+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 ### Added
+- CLI help: `--help` or `-h`
 - CLI argument `-p` to open yesterday's log file
 - CLI argument `-p X` to open a log file `X` days in the past
+- CLI argument `--config` to open the user config file
+- CLI argument `--dir` to open the logfile directory
 
 ## 0.4.1 - 2017-05-26
 ### Added